Fixing ddp command

+++ src/core/proto.h
@@ -158,8 +158,8 @@ int siril_fits_img_stats_float(float *array, long nx, long ny, int nullcheck,
 
 /****************** siril.h ******************/
 
-int threshlo(fits *fit, int level);
-int threshhi(fits *fit, int level);
+int threshlo(fits *fit, WORD level);
+int threshhi(fits *fit, WORD level);
 int nozero(fits *fit, WORD level);
 int unsharp(fits*, double sigma, double mult, gboolean verbose);
 double entropy(fits *fit, int layer, rectangle *area, imstats *opt_stats);
